Add SD-Ultra endpoint and request logging; small README updates (#17)
* Add SD-Ultra endpoint and request logging; small README updates

Renamed `stable-diffusion-generate-image` to
`stable-diffusion-generate-image-core` and aliased the old endpoint to
it, to contrast with ultra.

Added a feature to log request and response, for successful generation,
and optionally also for failed.

Small README updates

* Implement suggestions: Warn model against always using SD Ultra, remove default generic endpoint.

Also added language to the description of the negative prompt to
encourage its use. In my quick testing this resulted in better quality
images with complex prompts.

* Clarify that Core should be the default in most cases where the user
doesn't have a specific preference

* Update src/tools/generateImageCore.ts

---------

Co-authored-by: Tadas Antanavicius <tadas@tadasant.com>